ENH: maybe fixes label comparison
